incoming: introduce process_measures_for_sack
This adds a new method process_measures_for_sack to incoming storage driver. It
allows to read an entire sack for new measure rather than individual metrics.
This avoids doing 2 listing to process new metrics and only does one to load
the measures.

The process_new_measures(metrics) is kept for the refresh_metric() use case for
now. Some refactoring might be possible after this patch.

The S3 storage driver is a bit modified to store the incoming measures in
<sack>/<metric>/ rather than <sack><metric>/ so it's easier to list incoming
sacks and metrics.