chef: merge refresh_metric and process_new_measures
This simplifies the code base by just allowing to process new measures sack per
sack with the existing process_new_measures_for_sack() or to use the new
refresh_metrics() who takes a list of metrics. Since the list of metric must be
known, there's no need to use list_metric_with_measures_to_process() anymore,
so it can be removed.

Since some of the tests (test_rest) need to process all sacks now, reduce the
number of sack to a small number in order for each test not to take 10s.